converted the app to use es6 and added eslint
es6 introduces several syntax features that allow us to write cleaner code such
as  support for proper classes, string interpolations, proper import/export
statements etc.

Under node this requires the use of babel, which does create a bit of an extra
burden in that files need to be build.

This commit introduces 2 new run commands:
- `build` which runs babel to convert the es6 code to something supported by node,
the output is stored in the /dist folder
- `dev` which watches the source dir and autobuilds on changes, to be used
during development
- `lint` runs eslint on the source to find possible syntax errors and other issues